requirements.txt
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coverage==7.2.7
flake8==6.0.0
handler==1.3.0
idna==3.4
mccabe==0.7.0
Pillow==10.0.0
pony==0.7.16
psycopg2==2.9.7
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
requests==2.31.0
six==1.16.0
urllib3==2.0.2
vk-api==11.9.9
